Refactored plugin warnings.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@90009 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfsound/freeze.c
@@ -217,8 +217,8 @@ void LoadStateUnknown(SPUFreeze_t * pF)
    s_chan[i].bNew=0;
    s_chan[i].bStop=0;
    s_chan[i].ADSR.lVolume=0;
-   s_chan[i].pLoop=(unsigned char *)((int)spuMemC+4096);
-   s_chan[i].pStart=(unsigned char *)((int)spuMemC+4096);
+   s_chan[i].pLoop=(unsigned char *)(spuMemC+4096);
+   s_chan[i].pStart=(unsigned char *)(spuMemC+4096);
    s_chan[i].iMute=0;
    s_chan[i].iIrqDone=0;
   }
